(DONE) Percentile
Calculating percentile now works, it takes all the chicken currently in the game, adds their stats together and sorts them by this value. We then find the index of the chicken we want to learn the percentile of and multiply this value by 100 before dividing by the total number of chickens. The Debug.Log will show the ID of each chicken as well as the total stat value in order, it then outputs the percentile of the chicken with ID "0".