Breeding
Finished the breeding algorithm so it now correctly selects stats from
both parents and passes them down into a baby Chicken. There is one
variable (breedingPotency) that isn't being calculated yet since that
was going to be based on number of time the chicken has
bred/trained/raced already, so that would be added in later. For now the
variable just equals 0.