Small Additions
Player chickens now begin with a Breeding Potency of 10, when they breed
it increases by 0.5 and when they train it decreases by 0.2. During
breeding, after a stat has been selected we do another roll to check for
a mutation. At the moment the mutation is just that it ignores the
parents stats are rolls a random number between 0 and 25.